Allow deletion of CouchDB IDs which are not present in SQL Server.
This is a little dangerous (as it could indicate a problem with the
database state), but there are a very large number of CouchDB records
(over 30%) which are uninteresting for SQL Server users. The other
option would be to store their IDs anyway, achieving greater validitiy
but with a performance and storage penalty.